Extract common part of query events
